% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cba}
\alias{cba}
\title{Calculate Cronbachs Alpha from supplied variables.}
\usage{
cba(x)
}
\arguments{
\item{x}{A data-frame or matrix of numerical values where rows are across-items within-respondent observation vectors, and columns are within-item across-respondents observation vectors.}
}
\value{
Cronbachs Alpha for the sumscore of supplied variables.
}
\description{
Calculates Cronbachs Alpha, a very commonly used index for assessing the reliability / internal consistency of a sumscore. Often interpreted as the mean correlation across all possible split-half alternate forms of the test.
}
\note{
Missing values are treated by passing \code{na.rm = TRUE} to the \code{var} function call.

Be aware that this function does not issue a warning if there are negative correlations between variables in the supplied data-set.
}
\examples{
data <- expand.table(LSAT7)
cba(data)
}
