% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmat.R
\name{cmat}
\alias{cmat}
\title{Confusion Matrix for a test.}
\usage{
cmat(score, se, cut)
}
\arguments{
\item{score}{A vector of test-scores.}

\item{se}{A vector of standard errors associated with the test-scores.}

\item{cut}{A value indicating the cutoff-score of the test.}
}
\value{
A confusion matrix indicating the true- and false positive and negative rates of a test, given a specific cut-score and under the assumption that sampling errors distribute normally around true-values.
}
\description{
Given a set of scores, their standard errors of measurement and the cutoff value of a test, calculate a confusion matrix indicating true and false positives and negatives under the assumption that sampling errors distribute normally around true-values.
}
\references{
Rudner, L. (2005). Expected Classification Accuracy. Practical Assessment, Research and Evaluation. 10(13), p. 1-5.
}
