% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmat.R
\name{dpstats}
\alias{dpstats}
\title{Diagnostic Performance Statistics.}
\usage{
dpstats(cmat = NULL)
}
\arguments{
\item{cmat}{A confusion matrix of true and false positives and negatives, organized as \code{matrix(c(tp, fp, tn, fn), nrow = 2)}.}
}
\description{
Given a confusion matrix of the sort \code{matrix(c(tp, fp, tn, fn), nrow = 2)}, calculate common diagnostic performance statistics.
}
