% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsea.R
\name{RMSEA}
\alias{RMSEA}
\title{Root Mean Squared Error of Approximation (RMSEA).}
\usage{
RMSEA(chi2, df, n)
}
\arguments{
\item{chi2}{Chi-squared.}

\item{df}{The Chi-squared statistics Degrees of Freedom.}

\item{ss}{The sample size.}
}
\value{
A single value: The RMSEA statistic.
}
\description{
Calculates the Root Mean Squared Error of Approximation (RMSEA) by way of the Chi-squared statistic, its Degrees of Freedom, and the sample size.
}
