% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacPlot.R
\name{cacPlot}
\alias{cacPlot}
\title{Plotting tool for gauging the classification accuracy or consistency of classifications based on IRT ability estimates, where classification accuracy and consistency indices are computed by means of the \code{cacIRT} package (Lathrop, 2015).}
\usage{
cacPlot(x, ablty = NULL, ablty.se = NULL, stat = "ca",
  mdl = "Rasch", cutoff = 0, ci = TRUE, cSEM = FALSE,
  xRng = c(-3, 3), yRng = c(0, 1), grid = TRUE, lbls = TRUE,
  lgd = TRUE, rel.wdth = c(7, 1), colorblindFriendly = FALSE)
}
\arguments{
\item{x}{A data.frame or matrix with rows representing respondents and columns representing items, or a mirt-model object of class "SingleGroupClass".}

\item{ablty}{A vector of ability estimates. Required specification of standard error in \code{ablty.se}.}

\item{ablty.se}{A vector of standard errors of estimates corresponding to the values in the \code{ablty} vector.}

\item{stat}{A character-value indicating whether to color-code observations with respect to their expected consistency or accuracy. Permissible values are "c", "cc", "consistency" or "Consistency" for expected classification consistency, and "a", "ca", "accuracy" or "Accuracy" for expected classification accuracy.}

\item{mdl}{If a dataset was supplied as input, specifies which model to fit to the data by way of the \code{mirt} package (Chalmers, 2012). See \code{?mirt} for options.}

\item{cutoff}{The cutoff value relative to which expected classification consistency or accuracy for observations are to be calculated and illustrated.}

\item{ci}{Plot confidence intervals around each observation point?}

\item{cSEM}{Plot the conditional standard errors of the estimates?}

\item{xRng}{The range of the plotted x-axis.}

\item{yRng}{The range of the plotted y-axis.}

\item{grid}{Include a grid in the plot?}

\item{lbls}{Include labels in the plot?}

\item{rel.wdth}{The relative widths of the main plot and the color gradient legend.}

\item{colorblindFriendly}{Make gradient colorblind friendly?}
}
\value{
A graph plotting observations with color gradients indicating expected classification consistency and accuracy relative to a defined cutoff point.
}
\description{
Plotting tool for gauging the classification accuracy or consistency of classifications based on IRT ability estimates, where classification accuracy and consistency indices are computed by means of the \code{cacIRT} package (Lathrop, 2015).
}
\examples{
data <- expand.table(LSAT7[2:31, ])
cacPlot(data, stat = "c")
}
\references{
R. Philip Chalmers (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29.

Quinn N. Lathrop (2015). cacIRT: Classification Accuracy and Consistency under Item Response Theory. R package version 1.4.
}
