% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacPlot.R
\name{cacPlot}
\alias{cacPlot}
\title{Classification Accuracy and Consistency Plotting.}
\usage{
cacPlot(x = NULL, ablty = NULL, ablty.se = NULL, pop.dist = c(0,
  1), cutoff = 0, stat = "ca", ci = TRUE, cSEM = FALSE,
  xRng = c(-3, 3), yRng = c(0, 0.5), grid = TRUE, lbls = TRUE,
  rel.wdth = c(7, 1), mdl = "Rasch", ablty.est = "ML",
  colorblindFriendly = FALSE)
}
\arguments{
\item{x}{A data.frame or matrix with rows representing respondents and columns representing items, or a mirt-model object of class "SingleGroupClass".}

\item{ablty}{A vector of ability estimates. Requires specification of standard error in \code{ablty.se}.}

\item{ablty.se}{A vector of standard errors of estimates corresponding to the values in the \code{ablty} vector. Must be supplied if raw ability estimates are to be used for plotting.}

\item{pop.dist}{Population distribution. Specifies the mean and standard deviation of the population distribution along which individual Theta values are plotted.}

\item{cutoff}{The cutoff value relative to which expected classification consistency or accuracy for observations are to be calculated and illustrated.}

\item{stat}{A character-value indicating whether to color-code observations with respect to their expected consistency or accuracy. Permissible values are "c", "cc", "consistency" or "Consistency" for expected classification consistency, and "a", "ca", "accuracy" or "Accuracy" for expected classification accuracy.}

\item{ci}{Plot confidence intervals around each observation point?}

\item{cSEM}{Plot the conditional standard errors of the estimates?}

\item{xRng}{The range of the plotted x-axis.}

\item{yRng}{The range of the plotted y-axis.}

\item{grid}{Include a grid in the plot?}

\item{lbls}{Include labels in the plot?}

\item{rel.wdth}{The relative widths of the main plot and the color gradient legend.}

\item{mdl}{If a dataset was supplied as input, specifies which model to fit to the data by way of the \code{mirt} package (Chalmers, 2012). See \code{?mirt} for options.}

\item{ablty.est}{A character value specifying which estimator to use for estimating ability from data. Default is maximum likelihood ("ML"). See \code{?mirt} for options.}

\item{colorblindFriendly}{Make gradient color-blind friendly?}
}
\value{
A graph plotting observations with color gradients indicating expected classification consistency and accuracy relative to a defined cutoff point.
}
\description{
Plotting tool for gauging the classification accuracy or consistency of classifications based on IRT ability estimates, where classification accuracy and consistency indices are computed by means of the \code{cacIRT} package (Lathrop, 2015).

Classification accuracy (CA) refers to the rate of correct classification based on observed performances. Classification consistency (CC) refers to the rate at which examinees are placed in the same category across administrations of equivalent tests, regardless of whether the classification is accurate. Both of these values can range between 1 (perfect accuracy / consistency) and .5 (perfectly inaccurate / inconsistent; Lathrop and Cheng, 2014).

The expected classification accuracy and consistency indices are those of the "Rudner approach" (Rudner, 2001, 2005). The expected classification accuracy index gives the probability of an examinee with a given actual-score attaining a specific observed score within some given interval (e.g., above or below some given cutoff score) on the actual-score scale.
}
\examples{
# Color-blind friendly plotting of classification consistency based on feeding
# cacPlot a dataset, where the cutoff-point is set to 0.5.
data <- expand.table(LSAT7[2:31, ])
cacPlot(data, stat = "c", cutoff = .5, colorblindFriendly = TRUE)

# Plotting of classification accuracy based on feeding cacPlot a mirt
# model-object along with plotting of the conditional standard error
# of measurement (cSEM), using the default cutoff of 0 (i.e., above
# or below average.
data <- expand.table(LSAT7[2:31, ])
LSAT7.mod <- mirt(data, model = 1, itemtype = "Rasch")
cacPlot(LSAT7.mod, stat = "a", cSEM = TRUE)

# Plotting of classification consistency based on feeding cacPlot a
# vector of raw ability estimates, and a vector with the standard
# errors corresponding to those ability estimates.
ability_estimates <- fscores(mirt(expand.table(LSAT7[2:31, ]), 1, "Rasch"),
    "ML", response.pattern = expand.table(LSAT7[2:31, ]))[, c("F1", "SE_F1")]
cacPlot(ablty = ability_estimates[, "F1"], ablty.se = ability_estimates[, "SE_F1"], stat = "c")

# Plotting of classification accuracy with several cutoff points
data <- expand.table(LSAT7[2:31, ])
LSAT7.mod <- mirt(data, model = 1, itemtype = "Rasch")
cacPlot(LSAT7.mod, stat = "a", cutoff = c(-.5, 0, .5))
}
\references{
R. Philip Chalmers (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. Journal of Statistical Software, 48(6), 1-29.

Quinn N. Lathrop (2015). cacIRT: Classification Accuracy and Consistency under Item Response Theory. R package version 1.4.

Lawrence M. Rudner (2001). Computing the Expected Proportions of Misclassified Examinees. Practical Assessment, Research & Evaluation., 7(14), 1-6.

Lawrence M. Rudner (2005). Expected Classification Accuracy. Practical Assessment, Research & Evaluation, 10(13), 1-5.
}
