% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cba}
\alias{cba}
\title{Calculate Cronbachs Alpha from supplied variables.}
\usage{
cba(x)
}
\arguments{
\item{x}{A data-frame or matrix of numerical values where rows are respondents and columns are items.}
}
\value{
Cronbachs Alpha for the sumscore of supplied variables.
}
\description{
Calculates Cronbachs Alpha, a very commonly used index for assessing the reliability / internal consistency of a sumscore
}
\note{
Missing values are treated by passing \code{na.rm = TRUE} to the \code{var} function.

Be aware that this function does not issue a warning if there are negative correlations between variables in the supplied data-set.
}
